requirements/base.txt
Django==5.0.4
djangorestframework==3.14.0
celery[redis]==5.3.4
django-celery-beat==2.6.0
psycopg[binary]==3.1.18
redis==4.6.0
gunicorn==21.2.0
requests==2.31.0
python-dateutil==2.8.2
pytz==2024.1
django-encrypted-model-fields==0.6.5
cryptography==42.0.5
python-decouple==3.8
sendgrid==6.11.0
Pillow==10.2.0
django-redis==5.4.0
